Remove UserService
